making segmentsFromString private[os] moving its tests to os package
